# Pool allocations made during inner loop for GJ.