This was low-hanging fruit identified by some benchmarks. For the associativity benchmark @yihozhang wrote up:

```
(datatype Math
  (Var i64)
  (Add Math Math)
)

(define start (Add (Var 1) (Add (Var 2) (Add (Var 3) (Add (Var 4) 
              (Add (Var 5) (Add (Var 6) (Add (Var 7) (Var 8)))))))))

(rewrite (Add x y) (Add y x))
(rewrite (Add x (Add y z)) (Add (Add x y) z))

(run 8)

(define end (Add (Var 8) (Add (Var 7) (Add (Var 6) (Add (Var 5) 
            (Add (Var 4) (Add (Var 3) (Add (Var 2) (Var 1)))))))))
(check (= start end))
```
Some anecdotal numbers from my workstation:

Before:
```
target/release/egg-smol assoc.egg  9.28s user 0.03s system 98% cpu 9.480 total
```
After:

```
target/release/egg-smol tests/assoc.egg  2.75s user 0.02s system 96% cpu 2.874 total
```